client & bank model

+++ Data/Clients/ClientRepository.cs
@@ -13,19 +13,7 @@ namespace APITransferencias.Models
     public class ClientRepository : IClientRepository
 
     {
-
-        //Estamos instanciando el driver que nos sirve para conectarnos a postgres, pasandole como argumento 
-        //en el constructor(seguramente recibe la conexión directamente al ser instanciada) nuestra variable 
-        //global que será asignada cuando instanciemos la clase, y usando el método ConnectionString de NpgsqlConnection.
-        //En dicho método estará el ConnectionString que definimos en Program en el método Connect de PostgreSqlConfiguration.
-        //En donde este tiene acceso y usa la configuración o las variables de conexión a la base de datos
-        protected NpgsqlConnection dbConnection()
-        {
-            return new NpgsqlConnection("Server = 127.0.0.1; Port = 5432; Database = interbanking_transfers; User Id = postgres; Password = hola1606;");
-        }
-
-
-        APIContext context;
+        private APIContext context;
 
         public ClientRepository(APIContext apiContext)
         {
@@ -40,34 +28,15 @@ namespace APITransferencias.Models
         public async Task<bool> createClient(Client clientParam)
         {
 
+            var request = context.Clients.Find(clientParam.cedula);
 
-            //var db = dbConnection();
+            if (request != null)
+            {
+                throw new Exception("Cliente ya existe");
+            }
+            context.Add(clientParam);
 
-            //string query = @$"
-
-            // INSERT INTO clientes (cedula, tipo_doc, nombre_apellido)
-            // VALUES(@cedula, @tipo_doc, @nombre_apellido)
-            // ";
-
-
-            //var filterClient = await getClientbyCedula(clientParam.cedula);
-
-            //if(filterClient != null)
-            //{
-            //    throw new Exception("Cliente ya se encuentra registrado");
-
-            //}
-
-
-            ////para hacer peticiones de tipo POST nos viene bien usar ExecuteAsync. Esto devuelve un int, si todo salió bien 
-            ////debemos revolver: response > 0. Ya que, devolverá 1 si se ejecuto correctamente la query o si al menos una 
-            ////fila fue afectada
-            //var response = await db.ExecuteAsync(query, new { clientParam.cedula, clientParam.tipo_doc, clientParam.nombre_apellido, 
-            //});
-
-            var client = context.Clients.Add(clientParam);
-
-            context.SaveChangesAsync();
+            await context.SaveChangesAsync();
 
             return true;
 
@@ -78,6 +47,11 @@ namespace APITransferencias.Models
          
             var client = context.Clients.Find(cedula);
 
+            if(client == null)
+            {
+                throw new Exception("Cliente no existe");
+            }
+
             var request = context.Clients.Remove(client);
 
             await context.SaveChangesAsync();
@@ -86,27 +60,31 @@ namespace APITransferencias.Models
         }
 
         public async Task<Client> getClientbyCedula(string cedula_cliente)
-        {
+        {   
             var request = context.Clients.Find(cedula_cliente);
+
+            if(request == null)
+            {
+                throw new Exception("Cliente no existe");
+            }
    
             return request;
         }
 
         public async Task<bool> updateClient(Client clientParam)
         {
-            //var db = dbConnection();
 
-            //string query = @$"
+            var client = context.Clients.Find(clientParam.cedula);
 
-            // UPDATE clientes
-            // SET tipo_doc = @tipo_doc,
-            // nombre_apellido = @nombre_apellido
-            // WHERE cedula = @cedula
-            // ";
+            if (client == null)
+            {
+                throw new Exception("Cliente no existe");
+            } 
+            
+            client.tipo_doc = clientParam.tipo_doc;
+            client.nombre_apellido = clientParam.nombre_apellido;
 
-            //var response = await db.ExecuteAsync(query, new { clientParam.cedula, clientParam.tipo_doc, clientParam.nombre_apellido });
-
-            //return response > 0;
+            await context.SaveChangesAsync();
 
             return true;
         }
